This gets most things rearranged so that the interface at least loads and it's more organized.
