+++ many-stage/app/common/frictionless_util.py
@@ -30,4 +30,3 @@ def make_readable_message(report):
             markdown_message += f"  - {message}\n"
 
     return markdown_message
-
