+++ many-stage/app/config.py
@@ -8,7 +8,13 @@ else:
 KEYS = ['openai_api_key', 'google_project', 'google_location', 'google_api_key']
 
 if not IN_COLAB:
-    with open("config.yaml", 'r') as stream:
+    # Get the directory of the current file
+    current_dir = os.path.dirname(os.path.abspath(__file__))
+
+    # Construct the path to config.yaml
+    config_path = os.path.join(current_dir, 'config.yaml')
+
+    with open(config_path, 'r') as stream:
         try:
             config = yaml.safe_load(stream)
         except yaml.YAMLError as exc:
